requirements.txt
asgiref==3.10.0
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
sqlparse==0.5.3
